Move notes related classes to separate package CIRC-742
