+++ src/main/java/org/folio/circulation/domain/notes/Note.java
@@ -1,9 +1,10 @@
-package org.folio.circulation.domain;
+package org.folio.circulation.domain.notes;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 import org.apache.commons.lang3.StringUtils;
+import org.folio.circulation.domain.NoteLink;
 import org.folio.circulation.support.JsonArrayHelper;
 
 import static org.folio.circulation.support.JsonPropertyFetcher.getArrayProperty;
